refactor: redesign of commands
The TUI is redefining what we need from CLI perspective. Also the
structure of subcommands is going to be irelevant.
