+++ cmd/kconf/cluster/item.go
@@ -1,4 +1,4 @@
-package clst
+package cluster
 
 import apiv1 "k8s.io/client-go/tools/clientcmd/api/v1"
