+++ cmd/kconf/context/context.go
@@ -1,60 +1,59 @@
-package usr
+package context
 
 import (
 	"fmt"
 
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/sn3d/kconf/cmd/kconf/usr/mod"
+	"github.com/sn3d/kconf/cmd/kconf/context/ls"
+	"github.com/sn3d/kconf/cmd/kconf/context/mv"
+	"github.com/sn3d/kconf/cmd/kconf/context/rm"
 	"github.com/sn3d/kconf/pkg/kconf"
 	"github.com/sn3d/kconf/pkg/tui/list"
 	"github.com/urfave/cli/v2"
 )
 
 var Cmd = &cli.Command{
-	Name:      "usr",
-	Usage:     "change user for context",
-	ArgsUsage: "[USER]",
+	Name:      "ctx",
+	Usage:     "context commands",
+	ArgsUsage: "[CONTEXT]",
 	Flags: []cli.Flag{
 		&cli.StringFlag{
 			Name:  "kubeconfig",
 			Usage: "path to kubeconfig from where export context",
 		},
-		&cli.StringFlag{
-			Name:    "context",
-			Aliases: []string{"c"},
-			Usage:   "context for which you want to change user",
-		},
 	},
 
 	Subcommands: []*cli.Command{
-		mod.Cmd,
+		mv.Cmd,
+		rm.Cmd,
+		ls.Cmd,
 	},
 
-	// main entry point
 	Action: func(cCtx *cli.Context) error {
-		user := cCtx.Args().First()
-		if user != "" {
-			return directChange(cCtx, user)
+		k8sContext := cCtx.Args().First()
+		if k8sContext != "" {
+			return directChange(cCtx, k8sContext)
 		} else {
 			return showTUI(cCtx)
 		}
-
 	},
 }
 
-func directChange(cCtx *cli.Context, user string) error {
+func directChange(cCtx *cli.Context, k8sContext string) error {
 
 	kc, err := kconf.Open(cCtx.String("kubeconfig"))
 	if err != nil {
 		return err
 	}
 
-	err = kc.ChangeUser(cCtx.String("context"), user)
+	kc.CurrentContext = k8sContext
+
+	err = kc.Save()
 	if err != nil {
 		return err
 	}
 
-	return kc.Save()
+	return nil
 }
 
 func showTUI(cCtx *cli.Context) error {
@@ -63,12 +62,15 @@ func showTUI(cCtx *cli.Context) error {
 		return err
 	}
 
-	model, err := NewModel(kc, cCtx.String("context"))
+	model, err := NewModel(kc)
 	if err != nil {
 		return err
 	}
 
 	m, err := tea.NewProgram(model, tea.WithAltScreen()).Run()
+	if err != nil {
+		return err
+	}
 
 	out := m.(Model)
 	switch msg := out.ExitMsg.(type) {
@@ -77,8 +79,9 @@ func showTUI(cCtx *cli.Context) error {
 			fmt.Printf("changes saved\n")
 		}
 	case list.PickedMsg:
-		fmt.Printf("user changed to %s\n", msg.Picked)
+		fmt.Printf("context changed to %s\n", msg.Picked)
 	}
 
 	return err
+
 }
