+++ cmd/kconf/context/item.go
@@ -1,4 +1,4 @@
-package ctx
+package context
 
 import (
 	"fmt"
