+++ main.go
@@ -2,15 +2,15 @@ package main
 
 import (
 	"fmt"
-	"github.com/sn3d/kconf/cmd/kconf/clst"
-	"github.com/sn3d/kconf/cmd/kconf/ctx"
-	"github.com/sn3d/kconf/cmd/kconf/ns"
-	"github.com/sn3d/kconf/cmd/kconf/usr"
 	"os"
+
+	"github.com/sn3d/kconf/cmd/kconf/cluster"
+	"github.com/sn3d/kconf/cmd/kconf/context"
+	"github.com/sn3d/kconf/cmd/kconf/ns"
+	"github.com/sn3d/kconf/cmd/kconf/user"
 
 	"github.com/sn3d/kconf/cmd/kconf/export"
 	"github.com/sn3d/kconf/cmd/kconf/imprt"
-	"github.com/sn3d/kconf/cmd/kconf/ls"
 	"github.com/sn3d/kconf/cmd/kconf/split"
 	"github.com/urfave/cli/v2"
 )
@@ -20,17 +20,17 @@ var version = "development"
 
 func main() {
 	app := &cli.App{
-		Name:    "kconf",
-		Version: version,
-		Usage:   "managing your kubeconfig",
+		Name:                 "kconf",
+		Version:              version,
+		Usage:                "managing your kubeconfig",
+		EnableBashCompletion: true,
 		Commands: []*cli.Command{
 			imprt.Cmd,
 			export.Cmd,
-			ctx.Cmd,
-			usr.Cmd,
-			clst.Cmd,
+			context.Cmd,
+			user.Cmd,
+			cluster.Cmd,
 			ns.Cmd,
-			ls.Cmd,
 			split.Cmd,
 		},
 	}
